==> Notebooks/02_raw_download.ipynb <==
import requests
#from bs4 import BeautifulSoup as bs
import pandas as pd
from io import StringIO
#import geopandas as gpd

# Part 2 Data Retrieval
Use the `requests` library and the USGS's API (https://earthquake.usgs.gov/fdsnws/event/1/) to retrieve information about all recorded earthquakes that occurred in Tennessee since 1900.
Answer the following questions:

#a list of all the catalogs available, if needed
cat_url = 'https://earthquake.usgs.gov/fdsnws/event/1/catalogs'
#The data dictionary for the dataset
#https://earthquake.usgs.gov/data/comcat/data-eventterms.php#magNst

#two catalogs seem useful
baseurl = 'https://earthquake.usgs.gov/fdsnws/event/1/query?format=csv&starttime=1900-01-01&endtime=2020-12-31&catalog='
catalog = ['ushis'
           , 'se'
          ]
tn_eq = pd.DataFrame()
for cat in catalog:
    url = baseurl + cat
    request = requests.get(url)
    temp_df = pd.read_csv(StringIO(request.text))
    tn_eq = pd.concat([temp_df, tn_eq])
    print(url)

#check that we have both catalogs
tn_eq.locationSource.unique()

#check
tn_eq.head()

#check
tn_eq.shape

#find any records that don't have a valid place
tn_eq[tn_eq['place'].isnull()]

#and drop
tn_eq = tn_eq.dropna(subset=['place'])
tn_eq.shape

#find any records that don't have a valid mag
tn_eq[tn_eq['mag'].isnull()]

#where are they?
tn_eq[tn_eq['mag'].isnull()]['place'].unique()

#How many in Tennessee?
tn_eq[(tn_eq['mag'].isnull()) & (tn_eq['place'] == 'Tennessee')]

#and drop
tn_eq = tn_eq.dropna(subset=['mag'])
tn_eq.shape

#filter for earthquake 'type'
#tn_eq.type.unique()
tn_eq = tn_eq[tn_eq['type'] == 'earthquake']
tn_eq.shape

#filter for tennessee 'place'
tn_eq = tn_eq[(tn_eq['place'].str.contains('Tenn'))
             | (tn_eq['place'].str.contains('TN'))
             ]
tn_eq.shape

tn_eq.columns

for col in tn_eq.columns:
    print(tn_eq[col].name)
    print(tn_eq[col].unique())

tn_eq.to_csv('../Data/tenn_eq.csv', index=False)
