==> Notebooks/03_correct_and_add_deaths.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import scipy.stats as stats
import statsmodels.api as sm
%matplotlib inline
pd.options.display.max_rows = 999
from ast import literal_eval
import re

eq_raw = pd.read_csv('../data/deadly_eq_raw.csv')

eq_raw.head()

np.where(eq_raw.applymap(lambda x: x == ''))

eq_nan = eq_raw.replace(r'^\s*$', np.NaN, regex=True)

eq_nan['em-dat total deaths'].nunique()

eq_nan.head()

np.where(eq_nan.applymap(lambda x: x == ''))

eq_nan[eq_nan['pde shaking deaths']>0]

eq_nan['other source deaths'].unique()

eq_nan['utsu total deaths'].unique()

def get_max_death(row):
    try:
        value = re.sub('\s\(.*',"",str(row))
        value = re.sub("[\(\[].*?[\)\]]", "",str(value))
        value = re.sub("\+", "",str(value))
        value = re.sub("\*", "",str(value))
        value = re.sub("\,", "",str(value))
        value = max([int(e) for e in value.split(' ')])
        return value
    except:
        return None


def clean_death(row):
    try:
        value = re.sub("[\(\[].*?[\)\]]", "",str(row))
        value = re.sub("\|", "",str(value))
        return value
    except:
        return None

eq_nan['em-dat total deaths'].nunique()

eq_nan['em-dat total deaths'] = [clean_death(row) for row in eq_nan['em-dat total deaths']]

eq_nan['em-dat total deaths'].nunique()

eq_nan['other source deaths'] = [get_max_death(row) for row in eq_nan['other source deaths']]

eq_nan['em-dat total deaths'].nunique()

orig_list = eq_nan['em-dat total deaths'].to_list()

orig_list

eq_nan['em-dat total deaths'] = pd.to_numeric(eq_nan['em-dat total deaths'],errors='coerce')

post_list = eq_nan['em-dat total deaths'].unique()

set([str(x) for x in post_list]).difference([str(float(x)) for x in orig_list])

set([str(float(x)) for x in orig_list]).difference([str(x) for x in post_list])

eq_nan['em-dat total deaths'].nunique()

[float(x) for x in eq_nan['em-dat total deaths'].unique() if x]

eq_nan['other source deaths'].unique()

[float(x) for x in eq_nan['other source deaths'].unique()]

eq_nan['deaths'] = eq_nan.max(axis=1)

eq_nan

eq_nan.to_csv('../Data/03_earthquakes_with_deaths.csv', index=False)
