==> Notebooks/04_convert_magnitude.ipynb <==
import pandas as pd

eq_table = pd.read_csv('../Data/03_earthquakes_with_deaths.csv', index_col=False)

eq_table['magnitude'].str.split(expand=True)[1].unique()

eq_table[['original_magnitude','original_magnitude_scale']] = eq_table['magnitude'].str.split(expand=True, n=1)

eq_table['original_magnitude'] = pd.to_numeric(eq_table['original_magnitude'])

# Conversion

The following is an example of how we can convert magnitudes to a similar scale; however, as seen in https://www.researchgate.net/publication/292771405_Globally_valid_relations_converting_Ms_mb_and_MJMA_to_Mw the scales are very similar

# ms_mag = eq_table[eq_table['original_magnitude_scale'].str.lower() == 'ms']

# def ms_to_mw(row):
#     result = float(row['original_magnitude'])
#     if 3.0 <= result <= 6.1:
#         result += 2.20
#     elif 6.2 <= result <= 8.0:
#         result -= 0.02
#     return result

# ms_mag['mag_in_mw'] = ms_mag.apply(ms_to_mw, axis=1)

# def ms_to_mw(row):
#     result = row['original_magnitude']
#     if 3.0 <= result <= 6.1:
#         result += 2.20
#     elif 6.2 <= result <= 8.0:
#         result -= 0.02
#     return result

eq_table.to_csv('../Data/04_deadly_eq_mag_cleaned.csv', index=False)
